FindBugs fixes
git-svn-id: https://svn.java.net/svn/glassfish~svn/trunk/main@55598 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5

+++ nucleus/admin/rest/rest-service/src/main/java/org/glassfish/admin/rest/generator/client/JavaClientGenerator.java
@@ -136,7 +136,9 @@ public class JavaClientGenerator extends ClientGenerator {
         try {
             File jarDir = Util.createTempDirectory();
             File jarFile = new File(jarDir, fileName);
-            jarFile.createNewFile();
+            if (!jarFile.createNewFile()) {
+                throw new RuntimeException("Unable to create new file"); //i18n
+            }
             jarFile.deleteOnExit();
             target = new JarOutputStream(new FileOutputStream(jarFile));
